[web] Detach component `AnimatedTab` from profile feature
